BaseService check for delete if has show_status change soft_delete mlModel Create fillable fix Translation fix add role helpers

+++ app/Services/BaseService.php
@@ -88,7 +88,12 @@ class BaseService
             $this->fileService()->deleteModelFile($model);
         }
 
-        $this->repository->destroy($id);
+        if ($model->hasShowStatus()) {
+            $this->repository->softDelete($id);
+        } else {
+            $this->repository->destroy($id);
+        }
+
     }
 
     /**
